+++ src/main/java/net/sf/jabref/importer/fileformat/JSONEntryParser.java
@@ -16,13 +16,17 @@
 
 package net.sf.jabref.importer.fileformat;
 
+import java.util.ArrayList;
+import java.util.List;
+
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.json.JSONArray;
 import org.json.JSONObject;
 
+import net.sf.jabref.bibtex.EntryTypes;
 import net.sf.jabref.model.entry.BibtexEntry;
-import net.sf.jabref.model.entry.BibtexEntryTypes;
+import net.sf.jabref.model.entry.MonthUtil;
 
 public class JSONEntryParser {
 
@@ -33,6 +37,12 @@ public class JSONEntryParser {
 
     }
 
+    /**
+     * Convert a JSONObject containing a bibJSON entry to a BibtexEntry
+     *
+     * @param bibJsonEntry The JSONObject to convert
+     * @return the converted BibtexEntry
+     */
     public BibtexEntry BibJSONtoBibtex(JSONObject bibJsonEntry) {
         // Fields that are directly accessible at the top level BibJson object
         String[] singleFieldStrings = {"year", "title", "abstract", "month"};
@@ -43,23 +53,20 @@ public class JSONEntryParser {
 
 
         BibtexEntry entry = new BibtexEntry();
-        entry.setType(BibtexEntryTypes.ARTICLE);
+        entry.setType(EntryTypes.getType("article"));
 
         // Authors
         if (bibJsonEntry.has("author")) {
             JSONArray authors = bibJsonEntry.getJSONArray("author");
-            StringBuffer sb = new StringBuffer();
+            List<String> authorList = new ArrayList<>();
             for (int i = 0; i < authors.length(); i++) {
                 if (authors.getJSONObject(i).has("name")) {
-                    sb.append(authors.getJSONObject(i).getString("name"));
-                    if (i < (authors.length() - 1)) {
-                        sb.append(" and ");
-                    }
+                    authorList.add(authors.getJSONObject(i).getString("name"));
                 } else {
                     LOGGER.info("Empty author name.");
                 }
             }
-            entry.setField("author", sb.toString());
+            entry.setField("author", String.join(" and ", authorList));
         } else {
             LOGGER.info("No author found.");
         }
@@ -103,16 +110,13 @@ public class JSONEntryParser {
         // Keywords
         if (bibJsonEntry.has("keywords")) {
             JSONArray keywords = bibJsonEntry.getJSONArray("keywords");
-            StringBuffer sb = new StringBuffer();
+            List<String> keywordList = new ArrayList<>();
             for (int i = 0; i < keywords.length(); i++) {
                 if (!keywords.isNull(i)) {
-                    sb.append(keywords.getString(i));
-                    if (i < (keywords.length() - 1)) {
-                        sb.append(", ");
-                    }
+                    keywordList.add(keywords.getString(i));
                 }
             }
-            entry.setField("keywords", sb.toString());
+            entry.putKeywords(keywordList);
         }
 
         // Identifiers
@@ -143,6 +147,102 @@ public class JSONEntryParser {
                     }
                 }
             }
+        }
+
+        return entry;
+    }
+
+    /**
+     * Convert a JSONObject obtained from http://api.springer.com/metadata/json to a BibtexEntry
+     *
+     * @param springerJsonEntry the JSONObject from search results
+     * @return the converted BibtexEntry
+     */
+    public static BibtexEntry SpringerJSONtoBibtex(JSONObject springerJsonEntry) {
+        // Fields that are directly accessible at the top level Json object
+        String[] singleFieldStrings = {"issn", "volume", "abstract", "doi", "title", "number",
+                "publisher"};
+
+        BibtexEntry entry = new BibtexEntry();
+        String nametype;
+
+        // Guess publication type
+        String isbn = springerJsonEntry.optString("isbn");
+        if (com.google.common.base.Strings.isNullOrEmpty(isbn)) {
+            // Probably article
+            entry.setType(EntryTypes.getType("article"));
+            nametype = "journal";
+        } else {
+            // Probably book chapter or from proceeding, go for book chapter
+            entry.setType(EntryTypes.getType("incollection"));
+            nametype = "booktitle";
+            entry.setField("isbn", isbn);
+        }
+
+        // Authors
+        if (springerJsonEntry.has("creators")) {
+            JSONArray authors = springerJsonEntry.getJSONArray("creators");
+            List<String> authorList = new ArrayList<>();
+            for (int i = 0; i < authors.length(); i++) {
+                if (authors.getJSONObject(i).has("creator")) {
+                    authorList.add(authors.getJSONObject(i).getString("creator"));
+                } else {
+                    LOGGER.info("Empty author name.");
+                }
+            }
+            entry.setField("author", String.join(" and ", authorList));
+        } else {
+            LOGGER.info("No author found.");
+        }
+
+        // Direct accessible fields
+        for (String field : singleFieldStrings) {
+            if (springerJsonEntry.has(field)) {
+                String text = springerJsonEntry.getString(field);
+                if (!text.isEmpty()) {
+                    entry.setField(field, text);
+                }
+            }
+        }
+
+        // Page numbers
+        if (springerJsonEntry.has("startingPage") && (springerJsonEntry.getString("startingPage").length() > 0)) {
+            if (springerJsonEntry.has("endPage") && (springerJsonEntry.getString("endPage").length() > 0)) {
+                entry.setField("pages",
+                        springerJsonEntry.getString("startingPage") + "--" + springerJsonEntry.getString("endPage"));
+            } else {
+                entry.setField("pages", springerJsonEntry.getString("startingPage"));
+            }
+        }
+
+        // Journal
+        if (springerJsonEntry.has("publicationName")) {
+            entry.setField(nametype, springerJsonEntry.getString("publicationName"));
+        }
+
+        // URL
+        if (springerJsonEntry.has("url")) {
+            JSONArray urlarray = springerJsonEntry.optJSONArray("url");
+            if (urlarray != null) {
+                entry.setField("url", urlarray.getJSONObject(0).optString("value"));
+            } else {
+                entry.setField("url", springerJsonEntry.optString("url"));
+            }
+        }
+
+        // Date
+        if (springerJsonEntry.has("publicationDate")) {
+            String date = springerJsonEntry.getString("publicationDate");
+            entry.setField("date", date); // For BibLatex
+            String dateparts[] = date.split("-");
+            entry.setField("year", dateparts[0]);
+            entry.setField("month", MonthUtil.getMonthByNumber(Integer.parseInt(dateparts[1])).bibtexFormat);
+        }
+
+        // Clean up abstract (often starting with Abstract)
+        String abstr = entry.getField("abstract");
+        if ((abstr != null) && abstr.startsWith("Abstract")) {
+            entry.setField("abstract", abstr.substring(8));
         }
 
         return entry;