+++ CHANGELOG.md
@@ -22,6 +22,7 @@ to [sourceforge feature requests](https://sourceforge.net/p/jabref/features/) by
  - Tab shows the minimal unique folder name substring if multiple database files share the same name
  - Added a page numbers integrity checker
  - Position and size of certain dialogs are stored and restored.
+ - Feature: Search Springer
  - Feature: Search DOAJ, Directory of Open Access Journals
  - Changes the old integrity check by improving the code base (+tests) and converting it to a simple issues table
  - Added combo box in MassSetFieldAction to simplify selecting the correct field name